requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
bcrypt==5.0.0
CacheControl==0.14.4
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
dataparser==0.0.2
dateparser==1.2.2
distro==1.9.0
ecdsa==0.19.1
fastapi==0.128.0
firebase_admin==7.1.0
-e git+https://github.com/T4t4n32/FlowTask.git@48cb1c5dbc1c3045c92338784131c9cb054ee27a#egg=flowtask
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.45.0
google-auth-httplib2==0.3.0
google-cloud-core==2.5.0
google-cloud-firestore==2.22.0
google-cloud-storage==3.7.0
google-crc32c==1.8.0
google-genai==1.56.0
google-generativeai==0.8.6
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
greenlet==3.3.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
Jinja2==3.1.6
librt==0.7.7
MarkupSafe==3.0.3
msgpack==1.1.2
mypy==1.19.1
mypy_extensions==1.1.0
passlib==1.7.4
pathspec==0.12.1
proto-plus==1.27.0
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
PyJWT==2.10.1
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.21
python-telegram-bot==22.5
pytz==2025.2
regex==2025.11.3
requests==2.32.5
rsa==4.9.1
ruff==0.14.10
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.45
starlette==0.50.0
tabulate==0.9.0
tenacity==9.1.2
tqdm==4.67.1
types-dateparser==1.2.2.20250809
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.2
uvicorn==0.40.0
websockets==15.0.1
